fix SelectTargetList & revert SetConstestedPVP changes
