+++ src/server/scripts/EasternKingdoms/BlackrockMountain/MoltenCore/boss_magmadar.cpp
@@ -94,10 +94,10 @@ public:
                 case EVENT_LAVA_BOMB_RANGED:
                 {
                     std::list<Unit*> targets;
-                    SelectTargetList(targets, [this](Unit* target)
+                    SelectTargetList(targets, 1, SelectTargetMethod::Random, 1, [this](Unit* target)
                     {
                         return target && target->IsPlayer() && target->GetDistance(me) > MELEE_TARGET_LOOKUP_DIST && target->GetDistance(me) < 100.0f;
-                    }, 1, SelectTargetMethod::Random);
+                    });
 
                     if (!targets.empty())
                     {
